link premium codes to parent codes
The fee calculation for some codes is expressed as a percentage of a parent code.  Previously, we had been using heuristics to determine the parent code to use when calculating the fee for a premium code.   These heuristics were good, but not perfect.   This PR adds the concept of a premium code to the model so that this relationship can be explicitly stated.

This PR also includes a significant redesign of the daily details tab.   This tab was getting busy and cumbersome with the addition of the fee calculations and premium relationships, so it was switched to a card design where you click on a card to edit it, with only a single card open at a time rather than having all inputs for all detail entries visible simultaneously.

The "Add a new day" button was removed.   The new design made it simpler to include the day input on all entries, making the "Add a new day" button fairly superflous.